JZ-90992 Fix vault agent accumulating memory by (1) making static secret cache key stable; (2) when health checking Vault secret for cache eviction, get the access token from the sink instead of using the one from the cached request.
A major change to limit how long to keep health checking for cache eviction. The "static_secret_retry_duration_in_minutes" (default 1 hour) and
"static_secret_max_retry_count" (default 72) in the config define the health check frequency up to how many times. The default values make it
check every hour for up to 3 days. Meaning that if Vault sever ever goes down, we have 72 hours to get it back. After 72 hours, the agent will
evict the cache no matter what. This is so that we don't have a goroutine and cache item that stays forever.

One fix to handle deleted secrets. When health check the secret and receive 404, it means the secret is removed, so go ahead and remove the cached secret.

In addition, made a function to check for static secret. This way the Vault agent will not try to use the static cache for dynamic secrets.